+++ pages/api/questions/import.ts
@@ -1,10 +1,7 @@
 import { NextApiRequest, NextApiResponse } from 'next'
 import { getServerSession } from 'next-auth'
 import { v4 as uuid } from 'uuid'
-import { parse } from 'csv-parse/sync'
-import formidable from 'formidable'
 import * as yup from 'yup'
-import fs from 'fs'
 
 // models
 import { getQuestionByQuestionText, createQuestion } from '@/models/questions'
@@ -14,6 +11,7 @@ import { addQuestionCategory } from '@/models/questionCategories'
 // utils
 import { authOptions } from '../auth/[...nextauth]'
 import { isSuperAdmin } from '@/utils/isSuperAdmin'
+import { getCsvFromRequest } from '@/utils/csv'
 
 export const config = {
   api: {
@@ -41,26 +39,7 @@ export default async function importQuestionsHandler(
   }
 
   try {
-    // Parse the form data
-    const form = formidable()
-    const uploadedResult = await form.parse(req)
-    const files = uploadedResult[1]
-
-    if (!files.csv) {
-      return res.status(400).json({ error: 'CSV file is required' })
-    }
-
-    const csvFile = files.csv[0]
-    if (!csvFile.mimetype?.includes('csv')) {
-      return res.status(400).json({ error: 'File must be a CSV' })
-    }
-
-    // Read and parse the CSV file
-    const csvContent = fs.readFileSync(csvFile.filepath, 'utf-8')
-    const records = parse(csvContent, {
-      columns: true,
-      skip_empty_lines: true,
-    })
+    const records = await getCsvFromRequest(req)
 
     const csvRowValidation = yup.object({
       question: yup.string().required('Question is required'),
@@ -128,9 +107,6 @@ export default async function importQuestionsHandler(
         }
       }
     }
-
-    // Clean up the uploaded file
-    fs.unlinkSync(csvFile.filepath)
 
     return res.status(200).json({
       questionsCount: newQuestions,
